+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - README.md: a chapter about "features";
 - README.md: changed the examples into the chapter "How to use"
 - benchmark.js: a new set of tests
+- index.js: error checking and handling
 
 ### Removed